Fix EIPs limit quota code
