+++ initial-setup/lambdas/src/load-limits-step.ts
@@ -1,6 +1,7 @@
 import { ServiceQuotas } from '@aws-pbmm/common-lambda/lib/aws/service-quotas';
 import { AcceleratorConfig } from '@aws-pbmm/common-lambda/lib/config';
 import { Account, getAccountId } from '@aws-pbmm/common-outputs/lib/accounts';
+import { Limit, LimitOutput } from '@aws-pbmm/common-outputs/lib/limits';
 import { STS } from '@aws-pbmm/common-lambda/lib/aws/sts';
 import { SecretsManager } from '@aws-pbmm/common-lambda/lib/aws/secrets-manager';
 
@@ -11,34 +12,16 @@ export interface LoadLimitsInput {
   assumeRoleName: string;
 }
 
-export interface LimitOutput {
-  accountKey: string;
-  limitKey: string;
-  serviceCode: string;
-  quotaCode: string;
-  value: number;
-}
-
 interface LimitCode {
   serviceCode: string;
   quotaCode: string;
   enabled: boolean;
 }
 
-// TODO Move this to common so we can use it from initial-setup/cdk
-enum Limit {
-  Ec2Eips = 'Amazon EC2/Number of EIPs',
-  VpcPerRegion = 'Amazon VPC/VPCs per Region',
-  VpcInterfaceEndpointsPerVpc = 'Amazon VPC/Interface VPC endpoints per VPC',
-  CloudFormationStackCount = 'AWS CloudFormation/Stack count',
-  CloudFormationStackSetPerAdmin = 'AWS CloudFormation/Stack sets per administrator account',
-  OrganizationsMaximumAccounts = 'AWS Organizations/Maximum accounts',
-}
-
 const LIMITS: { [limitKey: string]: LimitCode } = {
   [Limit.Ec2Eips]: {
     serviceCode: 'ec2',
-    quotaCode: 'L-D0B7243C',
+    quotaCode: 'L-0263D0A3',
     enabled: true,
   },
   [Limit.VpcPerRegion]: {
